% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_causal_recur.R
\name{plot.causal_recur_fit}
\alias{plot.causal_recur_fit}
\title{Plot method for causal_recur_fit}
\usage{
\method{plot}{causal_recur_fit}(x, ...)
}
\arguments{
\item{x}{An object of class \code{causal_recur_fit}.}

\item{...}{Additional arguments (ignored).}
}
\value{
Invisibly returns NULL.
}
\description{
For a \code{causal_recur_fit} object, prompt the user to run MCMC diagnostics.
}
