% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_causal_recur.R
\name{summary.causal_recur_fit}
\alias{summary.causal_recur_fit}
\title{Summary method for causal_recur_fit}
\usage{
\method{summary}{causal_recur_fit}(
  object,
  pars_to_report = c("beta_Y", "beta_A", "gamma_Y", "gamma_A"),
  ...
)
}
\arguments{
\item{object}{An object of class \code{causal_recur_fit}.}

\item{pars_to_report}{Character vector of parameter names or patterns to report
(default: c("beta_Y","beta_A","gamma_Y","gamma_A")).}

\item{...}{Additional arguments passed to \code{rstan::summary()}.}
}
\value{
Invisibly returns a data.frame of posterior means and 95\% intervals.
}
\description{
Summarize posterior estimates for a causal recurrent-event model fit.
}
