% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g_computation.R
\name{print.gcomp_out}
\alias{print.gcomp_out}
\title{Print method for g_computation output}
\usage{
\method{print}{gcomp_out}(x, ...)
}
\arguments{
\item{x}{An object of class \code{gcomp_out}, output of \code{g_computation()}.}

\item{...}{Further arguments (ignored).}
}
\value{
Invisibly returns a data.frame of the summary.
}
\description{
Print a summary table of causal contrasts delta(s, K+1) for a \code{gcomp_out} object.
}
