% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_summary_table.R
\name{print.result_summary_table}
\alias{print.result_summary_table}
\title{Print a result_summary_table}
\usage{
\method{print}{result_summary_table}(x, ...)
}
\arguments{
\item{x}{A \strong{result_summary_table} object.}

\item{...}{Ignored.}
}
\value{
\emph{x}, invisibly.
}
\description{
Formatted console/knitr/gt output of the two summary tables.
}
