% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result_summary_table.R
\name{summary.result_summary_table}
\alias{summary.result_summary_table}
\title{Summary for result_summary_table}
\usage{
\method{summary}{result_summary_table}(object, ...)
}
\arguments{
\item{object}{A \code{result_summary_table} object.}

\item{...}{Ignored.}
}
\value{
\emph{object}, invisibly.
}
\description{
Alias for \code{\link[=print.result_summary_table]{print.result_summary_table()}}.
}
