% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/propensity_score_diagnostics.R
\name{plot.ps_diag}
\alias{plot.ps_diag}
\title{Plot method for ps_diag}
\usage{
\method{plot}{ps_diag}(x, type = c("histogram", "density"), ...)
}
\arguments{
\item{x}{An object of class \code{ps_diag}.}

\item{type}{Character, either "histogram" or "density" (default: both available).}

\item{...}{Additional arguments passed to the underlying ggplot layers (e.g., adjust bins).}
}
\value{
Invisibly returns the ggplot object.
}
\description{
Plot propensity score diagnostics. User specifies \code{type} as "histogram" or "density".
}
